refactor: update `withReadPermission` imports
